+++ src/hash_test.cpp
@@ -1,4 +1,19 @@
-/* AFK (c) Alex Holloway 2013 */
+/* AFK
+ * Copyright (C) 2013, Alex Holloway.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see [http://www.gnu.org/licenses/].
+ */
 
 #include "afk.hpp"
 
@@ -12,15 +27,15 @@ void test_cellHash(void)
     std::cout << "Cell hash test" << std::endl;
     std::cout << "--------------" << std::endl;
 
-    for (long long x = -1; x <= 1; ++x)
+    for (int64_t x = -1; x <= 1; ++x)
     {
-        for (long long y = -1; y <= 1; ++y)
+        for (int64_t y = -1; y <= 1; ++y)
         {
-            for (long long z = -1; z <= 1; ++z)
+            for (int64_t z = -1; z <= 1; ++z)
             {
-                for (long long scale = 2; scale <= 8; scale *= 2)
+                for (int64_t scale = 2; scale <= 8; scale *= 2)
                 {
-                    AFK_Cell testCell = afk_cell(afk_vec4<long long>(x, y, z, scale));
+                    AFK_Cell testCell = afk_cell(afk_vec4<int64_t>(x, y, z, scale));
                     std::cout << testCell << " hashes to " << std::hex << hash_value(testCell) << std::endl;
                 }
             }
@@ -40,13 +55,13 @@ void test_tileHash(void)
     std::cout << "Tile hash test" << std::endl;
     std::cout << "--------------" << std::endl;
 
-    for (long long x = -1; x <= 1; ++x)
+    for (int64_t x = -1; x <= 1; ++x)
     {
-        for (long long z = -1; z <= 1; ++z)
+        for (int64_t z = -1; z <= 1; ++z)
         {
-            for (long long scale = 2; scale <= 16; scale *= 2)
+            for (int64_t scale = 2; scale <= 16; scale *= 2)
             {
-                AFK_Tile testTile = afk_tile(afk_vec3<long long>(x, z, scale));
+                AFK_Tile testTile = afk_tile(afk_vec3<int64_t>(x, z, scale));
                 std::cout << testTile << " hashes to " << std::hex << hash_value(testTile) << std::endl;
             }
         }
@@ -60,7 +75,7 @@ void test_rotate(void)
     std::cout << "Rotate test" << std::endl;
     std::cout << "-----------" << std::endl;
 
-    const long long test_values[] = {
+    const int64_t test_values[] = {
         1, -1, 0x040004000400040ll, -0x0400040000400040ll, 0x4000000000000000ll, -0x4000000000000000ll, 0
     };
 
@@ -70,7 +85,7 @@ void test_rotate(void)
         {
             std::cout << std::hex << test_values[i] <<
                 "<<" << std::dec << rot <<
-                " = " << std::hex << LROTATE_UNSIGNED((unsigned long long)test_values[i], rot) << std::endl;
+                " = " << std::hex << LROTATE_UNSIGNED((uint64_t)test_values[i], rot) << std::endl;
         }
     }
