+++ src/keyed_cell.cpp
@@ -1,4 +1,19 @@
-/* AFK (c) Alex Holloway 2013 */
+/* AFK
+ * Copyright (C) 2013, Alex Holloway.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see [http://www.gnu.org/licenses/].
+ */
 
 #include "afk.hpp"
 
@@ -44,7 +59,7 @@ AFK_KeyedCell afk_keyedCell(const AFK_KeyedCell& other)
     return cell;
 }
 
-AFK_KeyedCell afk_keyedCell(const AFK_Cell& _baseCell, long long _key)
+AFK_KeyedCell afk_keyedCell(const AFK_Cell& _baseCell, int64_t _key)
 {
     AFK_KeyedCell cell;
     cell.c = _baseCell;
@@ -52,7 +67,7 @@ AFK_KeyedCell afk_keyedCell(const AFK_Cell& _baseCell, long long _key)
     return cell;
 }
 
-AFK_KeyedCell afk_keyedCell(const Vec4<long long>& _coord, long long _key)
+AFK_KeyedCell afk_keyedCell(const Vec4<int64_t>& _coord, int64_t _key)
 {
     AFK_KeyedCell cell;
     cell.c = afk_cell(_coord);
