+++ src/tile.hpp
@@ -1,4 +1,19 @@
-/* AFK (c) Alex Holloway 2013 */
+/* AFK
+ * Copyright (C) 2013, Alex Holloway.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see [http://www.gnu.org/licenses/].
+ */
 
 #ifndef _AFK_TILE_H_
 #define _AFK_TILE_H_
@@ -16,7 +31,7 @@ class AFK_Tile
 public:
     AFK_Tile();
 
-    Vec3<long long> coord;
+    Vec3<int64_t> coord;
 
     bool operator==(const AFK_Tile& _tile) const;
     bool operator!=(const AFK_Tile& _tile) const;
@@ -40,10 +55,10 @@ public:
 };
 
 AFK_Tile afk_tile(const AFK_Tile& other);
-AFK_Tile afk_tile(const Vec3<long long>& _coord);
+AFK_Tile afk_tile(const Vec3<int64_t>& _coord);
 AFK_Tile afk_tile(const AFK_Cell& cell);
 
-AFK_Cell afk_cell(const AFK_Tile& tile, long long yCoord);
+AFK_Cell afk_cell(const AFK_Tile& tile, int64_t yCoord);
 
 /* For insertion into an unordered_map. */
 size_t hash_value(const AFK_Tile& tile);
